+++ code/test/sixsq/nuvla/server/resources/nuvlabox_0_lifecycle_test.clj
@@ -395,6 +395,8 @@
               (ltu/is-operation-absent :activate)
               (ltu/is-operation-present :commission)
               (ltu/is-operation-present :decommission)
+              (ltu/is-operation-present :check-api)
+              (ltu/is-operation-present :reboot)
               (ltu/is-key-value :state "COMMISSIONED"))
 
           ;; check that services exist
@@ -431,6 +433,47 @@
 
                 )))
 
+          ;; check custom operations
+          ;;
+          (let [check-api (-> session
+                              (request nuvlabox-url)
+                              (ltu/body->edn)
+                              (ltu/is-status 200)
+                              (ltu/is-operation-present :edit)
+                              (ltu/is-operation-absent :delete)
+                              (ltu/is-operation-absent :activate)
+                              (ltu/is-operation-present :commission)
+                              (ltu/is-operation-present :decommission)
+                              (ltu/is-operation-present :check-api)
+                              (ltu/is-operation-present :reboot)
+                              (ltu/is-key-value :state "COMMISSIONED")
+                              (ltu/get-op-url :check-api))
+
+                reboot    (-> session
+                              (request nuvlabox-url)
+                              (ltu/body->edn)
+                              (ltu/is-status 200)
+                              (ltu/is-operation-present :edit)
+                              (ltu/is-operation-absent :delete)
+                              (ltu/is-operation-absent :activate)
+                              (ltu/is-operation-present :commission)
+                              (ltu/is-operation-present :decommission)
+                              (ltu/is-operation-present :check-api)
+                              (ltu/is-operation-present :reboot)
+                              (ltu/is-key-value :state "COMMISSIONED")
+                              (ltu/get-op-url :reboot))]
+
+            ;; check-api action
+            (-> session
+                (request check-api)
+                (ltu/body->edn)
+                (ltu/is-status 202))
+
+            ;; reboot action
+            (-> session
+                (request reboot)
+                (ltu/body->edn)
+                (ltu/is-status 202)))
 
           ;; second commissioning of the resource (with swarm credentials)
           (-> session